+++ src/main/java/com/vtence/molecule/servers/UndertowServer.java
@@ -121,6 +121,7 @@ public class UndertowServer implements Server {
         }
 
         private void setRequestInfo(Request request, HttpServerExchange exchange) {
+            request.serverHost(host);
             request.uri(exchange.getRequestURI() + queryComponent(exchange));
             request.path(exchange.getRequestPath());
             request.query(exchange.getQueryString());
@@ -129,7 +130,6 @@ public class UndertowServer implements Server {
             request.remoteHost(exchange.getSourceAddress().getHostName());
             request.timestamp(exchange.getRequestStartTime());
             request.scheme(exchange.getRequestScheme());
-            request.hostname(exchange.getHostName());
             request.port(exchange.getHostPort());
             request.protocol(exchange.getProtocol().toString());
             request.secure(exchange.getConnection().getSslSessionInfo() != null);